Gravity Function
Added gravity function, can short click space for small hop and hold for
big hop.
BUGS: player can fall through floor if they are going fast enough
because they will skip the platform location entirely.

+++ BaseGame99/GameObject.h
@@ -5,10 +5,10 @@ public:
 	GameObject();
 	~GameObject();
 
-	int x_location;
-	int y_location;
-	int x_speed;
-	int y_speed;
+	float x_location;
+	float y_location;
+	float x_speed;
+	float y_speed;
 
 	ALLEGRO_BITMAP* bmp;
